backend/services/ml_summary.py
import google.generativeai as genai
import json
from dotenv import load_dotenv
import os

load_dotenv()
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

genai.configure(api_key=GOOGLE_API_KEY)

try:
    genai.configure(api_key=GOOGLE_API_KEY)
except Exception as e:
    print(f"Error configuring Google AI: {e}")

def get_ai_summary(analysis_json: str) -> str:
    """
    Sends the analysis JSON to the AI model with a specific prompt
    and returns the generated user-friendly summary.
    """
    # The master prompt designed to get the best response from the AI
    prompt = f"""
    Role and Goal:
    You are "Forma," an encouraging and knowledgeable AI fitness coach. Your goal is to provide a motivational summary of a user's workout based on a JSON analysis report. You must be positive and constructive, focusing on celebrating progress while offering clear, simple, and actionable advice.

    Task:
    Analyze the following JSON data from a user's exercise video. Then, generate a summary in Markdown format that includes the following sections:
    1. A Quick Greeting: A short, upbeat opening.
    2. Your Workout Snapshot: A summary of the key performance metrics (reps, duration, intensity).
    3. Form Deep Dive: For each item in the `form_feedback` array, explain the issue in simple terms, describe *why* it's important to fix, and provide one clear, actionable tip to improve it. If there is no form feedback, congratulate the user on their excellent form.
    4. Coach's Final Word: A concluding paragraph that reinforces their effort and gives them something to focus on for next time.

    Critical Instructions:
    - Tone: Be motivational, not critical. Use words like "Let's focus on..." instead of "You did this wrong."
    - Simplicity: Avoid technical jargon. Explain concepts in a way anyone can understand.
    - Actionable Advice: The tips must be practical.
    - Handling No Reps: If `total_repetitions` is 0, the feedback should be gentle. Explain that no valid reps were detected and suggest checking the camera angle or making sure the correct exercise was selected. Do not analyze the form feedback in this case.

    Here is the JSON data:
    {analysis_json}
    """

    try:
        model = genai.GenerativeModel('gemini-2.5-flash')
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        print(f"An error occurred while calling the AI model: {e}")
        return "Sorry, there was an issue generating your workout summary. Please try again later."