requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy==2.0.34
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
mysql-connector-python==9.0.0
opencv-python>=4.10
mediapipe>=0.10
numpy>=1.26
scipy>=1.11
streamlit>=1.39
pydantic-settings>=2.0
google-generativeai>=0.3